Initial commit.
This is the first commit to the public repo after a some time of
in-house development, during mid-2016 and early 2017.

The name of a project is a dedication to german mathematician Georg
Friedrich Bernhard Riemann, who investigated geometry deeply.

This rieman project also deals a lot with geometry, but much more simple
and has nothing to do with hyperspaces of any kind.