Fixed startup without number of desktops property set.
The typical situation is a start of rieman before the WM, in some
start-up script.

Before, missing property lead to number of desktops set zero, what
caused trap error due to division by zero in rendering code.

After the fix, if property is not found, number of desktops is set to 1.