Prevented rendering deadlock with small windows.
The attempt to render a texture with 0x0 size (for an icon) lead to
complete stop of redrawing.  It looks like the reason may be some memory
corruption inside libcairo, but it is not clear at the moment.

The current fix is to simply avoid attempts to render something with at
least one zero dimension, as an obvious optimization.

The bug manifested itself clearly with Java (awt) applications that create 1x1
window on start and thus scale to 0x0.

Possible workaround is to disable icons rendering by setting in config:
<window_icon enable="false"/>