Added tracking of root window geometry.
The xcb->resolution was renamed to xcb->root_geom, as well as local
variables.

The notification mask for root window was updated to receive updates
for window size changes.  This happens when new monitor is connected
or moved via xrandr, for example.

The code calculating viewports are protected from possible zero result,
which could happen if root_geom.dim > desktop_geom.dim (and this should
not happen, as we properly update values, but it is not guaranteed
that root window does not exceed viewport, so fallback to single
viewport).