Refactoring: removed rie_event_state_t.
The event does not have state, but the pager does, so special state
structure is removed, and render() function interface was simplified.

The event name is not really needed to be saved, so output debug
messages instead.  The macros was added to get type of event instead
of passing the variable to store type.