Fixed issue with tiling.
The tile_adjust step is not actually necessary and may lead to incorrect
work in case of multiple configure events: the window geometry is not
yet updated in first event, and the resulting window size is wrong.
Also, it is not required to resize window once more on configure event.